server/requirements.txt
black==21.9b0
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
regex==2021.10.23
tomli==1.2.1
typing-extensions==3.10.0.2
Werkzeug==2.0.2
grpcio==1.44.0
grpcio-tools==1.44.0
psycopg2-binary==2.9.3
psycopg2